Fix borrow lifetime of this_cpu
The function create_task needs to be able to borrow this_cpu_mut, but
referring to this_cpu in one of the arguments of create_task results in
an intermediate borrow that outlives the function call.  Moving the call
outside of the function arguments separates the borrow lifetime to avoid
a conflict with create_task.

Signed-off-by: Jon Lange <jlange@microsoft.com>