cpu/apic: Add IPI support to local APIC handling
This change will perform IPI as required in response to a guest ICR
write.  Any interrupt vector requested by the guest remains within the
scope of SVSM control, so such interrupts are exempt from filtering
requirements.

The guest vCPU and SVSM vCPU contexts are assumed to have the same APIC
ID, and the host-provided APIC IDs are consumed without validation.  In
the future, the SVSM will require code to verify the uniqueness of APIC
IDs, as well as code to ensure that the assigned APIC IDs are consistent
with the topology defined by CPUID.

Signed-off-by: Jon Lange <jlange@microsoft.com>