cpu/idt: Remove the exception vector from the exception context
Removing the exception vector from the exceptoin context makes it
possible to capture an exception register context separately from the
vector that is invoking the handler - or even invoking a handler with a
NULL register context.  Both of these are required to support restricted
injection, where receipt of #HV establishes a register context without
identifying the vector being invoked, and where interrupt delivery may
occur outside of an exception handler.