sev: fix pvalidate page size issue
The PVALIDATE instruction takes a page size input indicating whether to
process the input as a 4 KB page or as a 2 MB page.  If a 2 MB page is
specified but the underlying virtual address mapping is a 4 KB mapping,
then the PVALIDATE will validate the 2 MB page described by the first 4
KB mapping and skip all of the remaining pages within the 2 MB virutal
span.  And if it turns out that the first 4 KB within the 2 MB page
describes a guest physical address that is not aligned to a 2 MB
boundary, the results are undefined.  Consequently, it is necessary to
ensure that any attempt to PVALIDATE a 2 MB page is only performed on an
address that is actually mapped as a 2 MB page.

Signed-off-by: Jon Lange <jlange@microsoft.com>