utils: introduce `ScopedRef`
`ScopedRef` and `ScopedMut` are designed to solve the problem of
managing lifetimes of references created from pointers.  Normally, when
a reference is created from a pointer (such as with `ptr::as_ref()`), it
is associated with the static lifetime, and as a result, the compiler is
unable to determine whether the reference will live long enough for its
intended use.  While functions like `ptr::as_ref()` can associate the
reference with a lifetime, the compiler cannot usefully use this
information to enfoce lifetime checks on pointers generated in this way
because although every reference can be bound to a lifetime, a reference
does not by itself own a lifetime, and without an owning lifetime, the
compiler has no way to know when the lifetime to which the reference is
bound goes out of scope.  The `ScopedRef` and `ScopedMut` objects solve
this by creating a new object every time a pointer is converted to a
reference, so there is an actual object with an associated lifetime that
the compiler can use to ensure that the reference remains valid.

Signed-off-by: Jon Lange <jlange@microsoft.com>