address_space: move the shared region out of the top PML4E
Every region needs to be representable with a base and top address.
However, the top address of the top PML4E does not fit into a 64-bit
number; it instead wraps around to zero.  The top PML4E should remain
unused to ensure that there is never a possibility of confronting
integer wraparound when calculating addresses for that region.