requests: handle interrupts when halted in the request loop
If the request loop determines that there is no active VMSA, the
processor enters a halt state.  This halt state can terminate if an
interrupt is received.  If that happens, the request loop will observe
that no change to the VMSA mapping has been requested.  In this case,
the request loop must return to the halt state rather than attempting to
process any requests.

Signed-off-by: Jon Lange <jlange@microsoft.com>