cpu/ipi: permit multicast IPI messages to hold atomic values
Atomic values are `Sync` and therefore are candidates for inclusion in
multicast IPI messages.  However, unless the value is copied back from
the shared buffer, it will not be visible to the sender.  Therefore, the
`IpiMessage` trait must expose a method that permits finalization of the
message send by copying back any atomic fields.

Signed-off-by: Jon Lange <jlange@microsoft.com>