task: fix some safety issues
Starting the scheduler is an unsafe operation because it relies on there
not being a current task, which is a guarantee the compiler cannot make
on its own.

Also, a function that returns a pointer is not necessarily unsafe.  The
operation of using the pointer is where the code becomes unsafe.

Signed-off-by: Jon Lange <jlange@microsoft.com>