igvm: use MADT for determining APIC IDs
The IGVM format specifies a mechanism for the host to inject an MADT
into the guest address space.  This MADT should be used to determine the
set of APIC IDs associated with the set of CPUs that will be started by
the guest.

Not all VMMs can support injection of an MADT through IGVM parameters.
When building for these systems, the IGVM file does not request an MADT,
and the MADT is obtained instead through QEMU firmware config.

Signed-off-by: Jon Lange <jlange@microsoft.com>