cpu: resolve safety issues with `PERCPU_AREAS`
The `PERCPU_AREAS` structure is not multi-thread safe because it
contains a `Vec` that is modified as additional CPUs are started.  This
would be safe when executing in a single-processor environment, but as
each processor is started, additional CPU state is pushed onto
`PERCPU_AREAS`, causing it to continue to change even after other
processors have started running.  In order for `PERCPU_AREAS` to be
sound, it must be fully populated before any additional processor can
begin running.

Signed-off-by: Jon Lange <jlange@microsoft.com>