requests: eliminate the request processing task
The request loop task and the request processing task both exist per
CPU, but only one can ever run at a time.  This was a useful design for
testing task wait queues, but it is not needed for task processing,
since the request processing logic can be called directly from the
request loop.

Signed-off-by: Jon Lange <jlange@microsoft.com>