requests: move vTPM handling into a separate task
Moving the vTPM protocol handler into a different task has the key
advantage of running the vTPM in an address space that is isolated from
the rest of the protocol handling logic.  In addition, keeping it in a
separate task means that it can be invoked from any other task, whether
that task is running in user mode or in kernel mode.  Having the vTPM
handler in a dedicated task restricts it to single-threaded operation,
but the vTPM is normally never invoked simultaneously from multiple
vCPUs so single-thread operation is not limiting.

Signed-off-by: Jon Lange <jlange@microsoft.com>