svsm: refactor guest firmware preparation
The preparation of guest firmware state can be divided into two separate
phases.

The first phase involves the configuration of guest memory,
including ensuring that page validation is in the state expected by
guest firmware, as well as ensuring that parameter information such as
the CPUID and secrets pages are placed into the guest address space.
This first phase has constant behavior for a given guest firmware (as
described in the boot parameters) regardless of what tasks will
ultimately be launched by the SVSM, and thus this will be a core part of
the kernel boot flow.  Much of this phase is also
platform-independent, and should not live in platform-specific code.

The second phase involves configuration of guest state that depends on
the operating mode of the SVSM.  This phase is left as a separate
operation that ultimately will be moved into the main guest processing
task.

Signed-off-by: Jon Lange <jlange@microsoft.com>