kernel: remove valid page bitmap from stage2
The second stage loader is guaranteed to validate the entirety of the
kernel memory region, either as part of the mapped kernel image or as
part of the kernel heap.  Consequently, there is no reason to construct
a valid bitmap in stage2 to pass to the kernel, because the kernel can
safely assume the correct state when it starts.  Furthermore, the kernel
does not attempt page validation before a new valid page bitmap can be
allocated from the heap, so there is no reason to rely on stage2 memory
to hold a temporary copy.

Signed-off-by: Jon Lange <jlange@microsoft.com>