vpu/vc: panic on a page not validated #VC
The SVSM is designed to validate all pages before attempting to use any
of them.  If a #VC exception occurs due to a page not being validated,
then the system should panic.  This must be checked before the system
attempts to emulate the faulting instruction, since the emulation
attempt may be incorrect if trying to read through any non-validated
pages.

Signed-off-by: Jon Lange <jlange@microsoft.com>