svsm: create transition page tables for AP startup
On platforms that use SIPI or stage1 to start secondary processors,
there must be a valid page table that can be used by the startup
environment that maps low memory as well as the shared global address
space so that a valid translation context exists during the transition
from 32-bit to 64-bit mode and into kernel execution before the correct
per-CPU page tables are loaded.  This startup page table can be
constructed in low memory adjacent to the SIPI stub where it can survive
long enough to be used for AP startup.

Signed-off-by: Jon Lange <jlange@microsoft.com>